requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlite3  # Built into Python
pydantic==2.5.0
openai==1.3.7
python-dotenv==1.0.0
structlog==23.2.0
uuid
datetime
